[FLINK-24507][table] Move TimestampStringUtils and document it
