+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/utils/TimestampStringUtils.java
@@ -16,15 +16,19 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.util;
+package org.apache.flink.table.planner.utils;
+
+import org.apache.flink.annotation.Internal;
 
 import org.apache.calcite.util.TimestampString;
 
 import java.time.LocalDateTime;
 
 /** Utility functions for calcite's {@link TimestampString}. */
+@Internal
 public class TimestampStringUtils {
 
+    /** Convert a {@link LocalDateTime} to a calcite's {@link TimestampString}. */
     public static TimestampString fromLocalDateTime(LocalDateTime ldt) {
         return new TimestampString(
                         ldt.getYear(),
@@ -36,24 +40,25 @@ public class TimestampStringUtils {
                 .withNanos(ldt.getNano());
     }
 
+    /** Convert a calcite's {@link TimestampString} to a {@link LocalDateTime}. */
     public static LocalDateTime toLocalDateTime(TimestampString timestampString) {
         final String v = timestampString.toString();
-        final int year = Integer.valueOf(v.substring(0, 4));
-        final int month = Integer.valueOf(v.substring(5, 7));
-        final int day = Integer.valueOf(v.substring(8, 10));
-        final int h = Integer.valueOf(v.substring(11, 13));
-        final int m = Integer.valueOf(v.substring(14, 16));
-        final int s = Integer.valueOf(v.substring(17, 19));
+        final int year = Integer.parseInt(v.substring(0, 4));
+        final int month = Integer.parseInt(v.substring(5, 7));
+        final int day = Integer.parseInt(v.substring(8, 10));
+        final int h = Integer.parseInt(v.substring(11, 13));
+        final int m = Integer.parseInt(v.substring(14, 16));
+        final int s = Integer.parseInt(v.substring(17, 19));
         final int nano = getNanosInSecond(v);
         return LocalDateTime.of(year, month, day, h, m, s, nano);
     }
 
     private static int getNanosInSecond(String v) {
-        switch (v.length()) {
-            case 19: // "1999-12-31 12:34:56"
-                return 0;
-            default: // "1999-12-31 12:34:56.789123456"
-                return Integer.valueOf(v.substring(20)) * (int) Math.pow(10, 9 - (v.length() - 20));
+        if (v.length() == 19) { // "1999-12-31 12:34:56"
+            return 0;
         }
+
+        // "1999-12-31 12:34:56.789123456"
+        return Integer.parseInt(v.substring(20)) * (int) Math.pow(10, 9 - (v.length() - 20));
     }
 }
